wbr-streamlit-bigquery/requirements.txt
streamlit
pandas
numpy
google-cloud-bigquery
plotly
streamlit-plotly-events
python-dotenv
pytest
